CSSTUDIO-1980: migrate select inputs to mui
